server/requirements.txt
argcomplete==3.1.4
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
dbus-python==1.3.2
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
filelock==3.13.1
Flask==3.0.2
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
netifaces==0.11.0
oauthlib==3.2.2
packaging==24.0
pipenv==2023.12.1
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyasyncore==1.0.2
pycurl==7.45.3
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyinotify==0.9.6
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu3
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
service-identity==24.1.0
setuptools==68.1.2
simplejson==3.19.2
six==1.16.0
systemd-python==235
Twisted==24.3.0
typing_extensions==4.10.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
userpath==1.9.1
virtualenv==20.25.0+ds
wadllib==1.3.6
Werkzeug==3.0.1
wheel==0.42.0
zope.interface==6.1
